[Refactor] Move shared types (BinaryOp, ComparisonOp, UnaryOp, Literal, Identifier, Label) to oneil_shared
These 6 types were defined identically in both oneil_ast and oneil_ir.
Now they live in oneil_shared with re-exports from both crates.
The resolver's identity-mapping functions are simplified to simple derefs.
Inlined also into the call sites trivial constructors.
