+++ src-rs/oneil_ir/src/parameter.rs
@@ -26,23 +26,7 @@ impl ParameterName {
     }
 }
 
-/// A label for a parameter.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Label(String);
-
-impl Label {
-    /// Creates a new label with the given name.
-    #[must_use]
-    pub const fn new(name: String) -> Self {
-        Self(name)
-    }
-
-    /// Returns the label as a string slice.
-    #[must_use]
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
+pub use oneil_shared::naming::Label;
 
 /// Represents a single parameter in an Oneil model.
 #[derive(Debug, Clone, PartialEq)]
