+++ src-rs/oneil_ir/src/reference.rs
@@ -1,24 +1,8 @@
 //! Reference types for identifiers, paths, and imports in Oneil.
 
+pub use oneil_shared::naming::Identifier;
+
 use std::path::{Path, PathBuf};
-
-/// An identifier for a variable, parameter, or other named entity in Oneil.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct Identifier(String);
-
-impl Identifier {
-    /// Creates a new identifier from a string or string-like type.
-    #[must_use]
-    pub const fn new(identifier: String) -> Self {
-        Self(identifier)
-    }
-
-    /// Returns the string value of this identifier.
-    #[must_use]
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
 
 /// A path to an Oneil model file.
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
