+++ src-rs/oneil_resolver/src/test/test_ir.rs
@@ -1,6 +1,6 @@
 use indexmap::IndexMap;
 
-use oneil_ir as ir;
+use oneil_ir::{self as ir, Literal};
 use oneil_shared::span::Span;
 
 /// Generates a span for testing purposes
@@ -19,7 +19,7 @@ pub fn reference_name(reference_name: &str) -> ir::ReferenceName {
 
 pub fn expr_literal_number(value: f64) -> ir::Expr {
     let span = unimportant_span();
-    ir::Expr::literal(span, ir::Literal::number(value))
+    ir::Expr::literal(span, Literal::Number(value))
 }
 
 pub fn empty_model() -> ir::Model {
