+++ src-rs/oneil_shared/src/expr_ops.rs
@@ -1,0 +1,65 @@
+//! Shared expression operator and literal types used by both AST and IR.
+
+/// Binary operators for mathematical and logical operations.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum BinaryOp {
+    /// Addition: `a + b`
+    Add,
+    /// Subtraction: `a - b`
+    Sub,
+    /// Escaped subtraction: `a -- b`
+    EscapedSub,
+    /// Multiplication: `a * b`
+    Mul,
+    /// Division: `a / b`
+    Div,
+    /// Escaped division: `a // b`
+    EscapedDiv,
+    /// Modulo: `a % b`
+    Mod,
+    /// Exponentiation: `a ^ b`
+    Pow,
+    /// Logical AND: `a && b`
+    And,
+    /// Logical OR: `a || b`
+    Or,
+    /// Min/max: `a | b`
+    MinMax,
+}
+
+/// Comparison operators for expressions.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum ComparisonOp {
+    /// Less than comparison: `a < b`
+    LessThan,
+    /// Less than or equal comparison: `a <= b`
+    LessThanEq,
+    /// Greater than comparison: `a > b`
+    GreaterThan,
+    /// Greater than or equal comparison: `a >= b`
+    GreaterThanEq,
+    /// Equality comparison: `a == b`
+    Eq,
+    /// Inequality comparison: `a != b`
+    NotEq,
+}
+
+/// Unary operators for single-operand operations.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum UnaryOp {
+    /// Negation: `-a`
+    Neg,
+    /// Logical NOT: `!a`
+    Not,
+}
+
+/// Literal values that can appear in expressions.
+#[derive(Debug, Clone, PartialEq)]
+pub enum Literal {
+    /// Numeric literal (floating-point).
+    Number(f64),
+    /// String literal.
+    String(String),
+    /// Boolean literal.
+    Boolean(bool),
+}
