+++ src-rs/oneil_shared/src/lib.rs
@@ -2,6 +2,8 @@
 //! Shared utilities for the Oneil programming language
 
 pub mod error;
+pub mod expr_ops;
 pub mod load_result;
+pub mod naming;
 pub mod partial;
 pub mod span;
