+++ src-rs/oneil_shared/src/naming.rs
@@ -1,0 +1,49 @@
+//! Shared naming types used by both AST and IR.
+
+/// An identifier for a variable, parameter, or other named entity in Oneil.
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct Identifier(String);
+
+impl Identifier {
+    /// Creates a new identifier with the given string value.
+    #[must_use]
+    pub const fn new(value: String) -> Self {
+        Self(value)
+    }
+
+    /// Returns the identifier as a string slice.
+    #[must_use]
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+}
+
+impl From<String> for Identifier {
+    fn from(value: String) -> Self {
+        Self::new(value)
+    }
+}
+
+/// A label in the Oneil language.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct Label(String);
+
+impl Label {
+    /// Creates a new label with the given string value.
+    #[must_use]
+    pub const fn new(value: String) -> Self {
+        Self(value)
+    }
+
+    /// Returns the label as a string slice.
+    #[must_use]
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+}
+
+impl From<String> for Label {
+    fn from(value: String) -> Self {
+        Self::new(value)
+    }
+}
